requirements.txt
rapidfuzz>=3.0
streamlit>=1.30
python-dotenv>=1.0
anthropic>=0.40  # optional — for LLM-enhanced mode
